handling clang Wall warnings as well

+++ common/plugin/PluginManager.C
@@ -544,7 +544,7 @@ PluginManager::GetPluginList(vector<pair<string,string> > &libs)
 #define PLUGIN_MAX(A,B) (((A) < (B)) ? (B) : (A))
 
             // Ignore it if it does not end in the correct extension
-            if (filename.length() < PLUGIN_MAX((1 + prefixLen),extLen) ||
+            if (filename.length() < (size_t)PLUGIN_MAX((1 + prefixLen),extLen) ||
                 !(filename.substr(filename.length()-extLen,extLen) == ext))
             {
                 continue;
